Create Eliminate Common Factor.cpp
This is from the recent CF's Div 2 problem, we need to find if a division will produce infinite number of decimals relative to a base number. A number will produce infinite fraction if the divisor (after simplifying the equation) have a factor that is not a factor of the base number.
At first I naively list all factors and then compare them to see if there is any member of factor q that is not factor of b. This got TLE. Then when the contest finished, I looked up other people's code, the solution is as simple as this:
If you gcd both number until the gcd is 1, that means we eliminate all the common factors between q and b. If the resulting q is 1, then all factors of q is also factor of b. If not, then it contains other factors which mean the fraction is infinite.